+++ __mocks__/gatsby-plugin-firebase.js
@@ -6,52 +6,7 @@ import {
   AuthConstants,
   DatabaseConstants,
 } from './gatsby-plugin-firebase/constants';
-
-class Auth {
-  static #instance = undefined;
-  #uuid = '';
-  #onAuthStateChangedObservers = [];
-
-  constructor() {
-    if (Auth.#instance) {
-      return Auth.#instance;
-    }
-
-    Auth.#instance = this;
-
-    this.#uuid = uuidv4();
-  }
-
-  get uuid() {
-    return this.#uuid;
-  }
-
-  get onAuthStateChangedObservers() {
-    return this.#onAuthStateChangedObservers;
-  }
-
-  dispose() {
-    this.#onAuthStateChangedObservers = [];
-  }
-
-  onAuthStateChanged(observer) {
-    this.#onAuthStateChangedObservers.push(observer);
-
-    return () => {
-      this.#onAuthStateChangedObservers = this.#onAuthStateChangedObservers.filter(
-        (observer) => observer !== observer,
-      );
-    };
-  }
-
-  async signInAnonymously() {
-    const user = AuthConstants.anonymousUser1;
-
-    this.#onAuthStateChangedObservers.forEach((observer) => observer(user));
-
-    return Promise.resolve(user);
-  }
-}
+import Auth from './gatsby-plugin-firebase/auth';
 
 class Database {
   static #instance = undefined;
